+++ mgl-webapp/src/main/java/com/organization/mvcproject/controller/GameController.java
@@ -13,8 +13,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
 
-import com.organization.mvcproject.model.Game;
-import com.organization.mvcproject.model.Review;
+import com.organization.mvcproject.api.model.Game;
+import com.organization.mvcproject.api.model.Review;
 import com.organization.mvcproject.service.GameService;
 
 @Controller
