--- mgl-webapp/src/main/java/com/organization/mvcproject/dao/MockGameDao.java
@@ -1,5 +1,0 @@
-package com.organization.mvcproject.dao;
-
-public class MockGameDao {
-
-}
