+++ mgl-webapp/src/main/java/com/organization/mvcproject/service/GameService.java
@@ -2,7 +2,7 @@ package com.organization.mvcproject.service;
 
 import java.util.List;
 
-import com.organization.mvcproject.model.Game;
+import com.organization.mvcproject.api.model.Game;
 
 public interface GameService {
