+++ mgl-webapp/src/main/java/com/organization/mvcproject/service/GameServiceImpl.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.organization.mvcproject.model.Game;
+import com.organization.mvcproject.api.model.Game;
 import com.organization.mvcproject.repository.GameDAOMock;
 
 @Service
