+++ mgl-webapp/src/test/java/com/organization/mvcproject/MGLTask1/service/Game_Service_ImplTest.java
@@ -25,8 +25,8 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.context.web.WebAppConfiguration;
 
+import com.organization.mvcproject.api.model.Game;
 import com.organization.mvcproject.config.MvcConfiguration;
-import com.organization.mvcproject.model.Game;
 import com.organization.mvcproject.service.GameService;
 
 @RunWith(JUnitPlatform.class)
